refactor: extract shared dispute helpers and deduplicate PDA derivations
Extract 4 identical helper functions from resolve_dispute.rs and
expire_dispute.rs into a shared dispute_helpers.rs module:
- validate_remaining_accounts_structure
- check_duplicate_arbiters
- process_arbiter_vote_pair
- process_worker_claim_pair

Replace 8 inline PublicKey.findProgramAddressSync calls in
autonomous/agent.ts with existing findClaimPda, findEscrowPda,
and findProtocolPda helpers.

Net reduction: ~126 lines of duplicated code.
